Improved identify_tracklet and added demonstration to notebook.
After conversation with Dave, it was clear that identify_tracklet would not
correctly identify tracklets with the same trksub as being different tracklets.
Dave said tot.itf is sorted by the trksub, at least when fully re-dumped.
Previously identify_tracklet only used the trksub to identify whether the next
line is or isn't the same tracklet as the prior. Now the UT date is also used,
allowing there to be up to a 24 hour gap (gap limit can be negotiated).